Update patch set 1
Patch Set 1:

I can't really judge this change.
For understanding: So OpenONUAC shall inherit the endpointManager interface and make use of its GetEndpoint() method. But is the endpointManager really started in the context of this adapter - maybe as part of the Kafka client or InterContainerProxy?
What is the adaptertype string to be used in getAdapterTopic() "openolt"?

I'm not sure, when this getAdapterTopic() method can be used. At the moment we set the adapterTopic only at initialization in main.go. From there I suppose this method is not reachable.

Is the usage of that method already working today? May be it's bound to specific subsystem versions (openolt,rwcore)?

Patch-set: 1
Reviewer: Gerrit User 1000666 <1000666@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>