Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is your intention to provide the 'getAdapterTopic' function to be used in any message sending to the OpenOlt Adapter(Instance)? This would theoretically also support some 'onTheFly' change of the partner OltAdapter(Instance). But I can't yet see, how this could be done by now (with existing library API's). For the moment we use Interprocess message sending to the OLT Adapter only for the transfer of OMCI messages. And here following API is used for sending: 	
> SendInterAdapterMessage(ctx context.Context,
> 		msg proto.Message,
> 		msgType ic.InterAdapterMessageType_Types,
> 		fromAdapter string,
> 		toAdapter string,
> 		toDeviceID string,
> 		proxyDeviceID string,
> 		messageID string) error
> I'm not sure, if here 'toAdapter' really is the same as 'toTopic'? Currently we use here the string as given in 'device.ProxyAddress.GetDeviceType()' at the time of 'adopt_device' (which at the moment is simply "openolt". Perhaps that is already sufficient to cope with different OLTAdapters (which would require to have the correct name in the GetDeviceType() function.
> If that is still something else than the toTopic, then we can only set the topic at the start of the adapter - which I still assume to be in main.go. Unfortunately at that time we also won't have any relevant device info to be used for setting the needed AdapterType.
> So in the end I still can't see how to make use of your extension :-(

The python version of openonu-adapter gets the openolt adapter's topic name as a command-line parameter. It was a recent change required to solve the problem when there are multiple OLT adapters in the same k8s deployment. 'device.ProxyAddress.GetDeviceType()' always returns 'openolt' but since we may have multiple OLT adapters in the same k8s deployment, it's not sufficient to use. That's why we needed to introduce olt-adapter's topic name as a command line parameter. It was a workaround. This API can be used to remove the need putting the olt-adapter's topic name as a command line parameter. However, just adding the API call is not sufficient. I wish we could see the whole implementation as a single patch and test it.

Patch-set: 1
Reviewer: Gerrit User 1000404 <1000404@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>