Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> @Himani: I downloaded your patch and applied golangci-lint using config .golangci.yml of https://gerrit.opencord.org/c/voltha-openonu-adapter-go/+/20219 on it. This produced errors. 
> Q: Was it your intention to fix all the SCA-issues with this patch or is it only a intermediate step?

There is one more review which I need to raise for the export issues. Almost all would be export issues. There are other sca errors introduced freshly with flows user story which I need to fix. Apart from that there are 7 errors which we need to ignore.

Patch-set: 3