Update patch set 5
Patch Set 5:

patch set 5 required to deliver the required flow reason state also in case internally nothing is really done (no change on the existing rules) (but flow setting is of course as required).
This does not solve the problem, that 'tech-profile-config-download-success' is the final state in case flow/techProfile configuration is done when the techProfile already exists for the onu adapter (as seen in one of the last POD testings [where the expected state is 'omci-flows-pushed']).

Patch-set: 5