Update patch set 4
Patch Set 4:

Seem you want to check the beavior with the updated robot test suite? Might be it is showing no error then, at least it did not in my test. But in the previous version in the test just the 1TCont1gem profile was used, and there were really problems. I still don't understand the flow of states observed there, but they show that some control is really still not working as intended. But looks like I have to reproduce locally and have to insert more debug info ...

Patch-set: 4