Update patch set 1
Patch Set 1:

(2 comments)

The implementation given here is not what I understood as request in the discussion, but if this was the request - then ok.
My understanding of the discussion was to just have the default setting of 'metricsDisabled' at start of the adapter, which should be able to be modified from some system call (globally and/or per group metrics and/or per standAlone metrics) at run time. This modification here just adds a switch that disables/enables all metrics for the life time of the adapter.
My understanding of some helm parameter for the metrics would then be to just set the initial state of metrics collection (if that is needed at all - it may possibly always start in the disabled state).
So this is more related on what the feature request is really about.

Patch-set: 1
Reviewer: Gerrit User 1000666 <1000666@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>