Update patch set 3
Patch Set 3:

Halfway patched also config.go to have the adapter running anyhow. (Difficult) Rebase would be required anyhow. 
In one of my test there was still a singular error with the critical sequence correction. Was not clearly understandable from the log, tried to make it a step more safe, but the sequence error does not appear that often to verify. Must be observed in ongoing regression tests.

Patch-set: 3