Update patch set 1
Patch Set 1:

General comment: I'd prefer to assign the omciTimeout variable to omci_cc and not to openonu. The context of this variable is functionally bound to the OMCI message transfer. This would also have the advantage of having this timer per ONU, once the system would be able to set ONU specific timeouts, which might make sense to reflect the different OMCI delay values that are possible depending on how the ONU's are detached. Accordingly I would remove the function parameter for all the calls to the omci_cc sendXXX functions and use this timeout value there directly - once the implementation really would use it.
Otherwise +1.

Patch-set: 1
Reviewer: Gerrit User 1000666 <1000666@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>