Update patch set 2
Patch Set 2: Code-Review+1

I tested with this patch and could not reproduce the issue from Friday anymore. But with the unmodified master of today I don't see this issue either! So +1 to see the results with the patch in CI.

Patch-set: 2
Reviewer: Gerrit User 1000671 <1000671@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=+1