Update patch set 7
Patch Set 7:

(3 comments)

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > This is a sample output from my testing with real ONUs (tested with Alpha XG-PON ONU)
> > 
> > $ voltctl device getextval onu_pon_optical_info 5247b2ab-6f7d-437e-bf01-62d67391e6e0
> > 
> >   POWER_FEED_VOLTAGE__VOLTS:      3
> >   RECEIVED_OPTICAL_POWER__dBu:    -16
> >   MEAN_OPTICAL_LAUNCH_POWER__dBu: 5
> >   LASER_BIAS_CURRENT__AMP:        0
> >   TEMPERATURE__Celsius:           61
> 
> - What is the resolution you have for PowerFeedVoltage? Rounded on V could be to course (guess 10mV resolution would be required at least).
> - Given power levels don't look correct to me, I'd suspect these levels are in dBm here .... I'd also think that the resolution should be at least 0,1dBm (compare discussion on metrics)
> - LaserBias unit I would expect to be in mA, resolution should perhaps be in 0,1 mA. The valie 0 looks quite strange to me.

I submitted this voltha-protos patch to provide decimal value resolution for the test results.
Regarding power level, even I think the the levels are in dB-milliwatt. 
The specification (check A3.39.5 in G.988 spec) says the OMCI value is "Decibel-microwatts, 2s complement, 0.002 dB resolution". I am slightly confused how to interpret this. When I multiple 0.002 to the OMCI value, this is rounded off value value I get(due to int32). Anyway the value seems to be in dB-milliwatt and not in dB-microwatt.

Patch-set: 7