Update patch set 2
Patch Set 2:

Now the job was failing like some of the previous before. But the test failure is not related to the modification done here, and maybe not even to the OnuAdapter. 
E.g. compare  'Sanity E2E Test for OLT/ONU on POD': It fails as if the expected default meter entry can't be found in ONOS. At that time no special activity was done within the OnuAdapter except for bringing the device into 'omci-flow-pushed' state. I think the default meter entry has nothing to do with the OnuAdapter processing, and changes done in this patch are only related to ONU upgrade procedures which are not tested here at all.
Perhaps someone else could have a closer look onto the issues here?

Patch-set: 2