Update patch set 10
Patch Set 10:

(2 comments)

The change is already acceptable - but I would welcome some further improvement -see code comments.
On the other hand it must be stated that the general treatment for the multi-UNI support is a bit questionable - not only related to this code change here. The problem specifically here is that there is just one event for the ONU now including all UNI port operStates in one variable. This might be valid for a straight-forward approach, but not generally. E.g. here the state indication is already delivered for all UNI's with the state change that comes from the first UNI port (independent from whether the unlocking for the other ports may fail or not). Generally I think it would be required to have different oper states per UNI, but to correctly support this further-going changes have to be introduced also on other code points (maybe also in different components), which would not be that simple. Maybe that should be addressed in some later independent story?

Patch-set: 10
Reviewer: Gerrit User 1000666 <1000666@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>