Update patch set 24
Patch Set 24:

(5 comments)

> Patch Set 24: Code-Review+1
> 
> (5 comments)
> 
> Had a rough view on processing.
> Some minor comments - perhaps at least check the inserted lib version.
> Concerning the Kv storage of the support/unsupport indication I'm still not sure if it is such a big advantage compared to just requesting at beginning (compared to the overhead needed for the KvStorage control). But this is anyway more a topic for Girish.
> Else: Good Job !

For the KV Store, storing the supported counters and then removing them from the next request would be overhead than advantage. Also it is only the last 2 counters which may not be supported as far we tested. so We can proceed with get request instead of storing in cache/kv store.

Patch-set: 24