Update patch set 1
Patch Set 1:

The main culprit was make([]bool,2) and then appending to the slice with true and false, so the resultant slice was coming up with 3 false values and 1 true value. Also, since redundant me will always give the same me instance id but always the pointer to managed entity will get changed, although managed entity contents would be same, hence for each redundant me, a different entry in the map was getting added. Correcting the make issue won't lead to this issue now, but anyways just wanted to take care of that as well.

Patch-set: 1