Update patch set 1
Patch Set 1:

I have to be honest that I do not understand the code changes. We must simplify this otherwise we risk losing ease of maintainability. We must serialize all the processing for a per {pon,onu,uni,tp} and I hope that can reduce a lot of complexity. For now I can merge the change as this helps fix the immediate regression we are seeing.

Patch-set: 1
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>