VOL-4774:openonuAdapterGo: Panic during scale test
RCA: In the scale tests, due to a rare timing complication between the go routine that act on deleting a device
and the go routine that does the collection metrics,the device resources are accessed by metric manager go routine
post deletion of the device, this is because of the timers that are started at the onu metric manager are not being
stopped during a deletion of device is called. This issue will be more predominant when the comm link b.w OLT adaptor
and the OLT agent is broken leading to increase in more wait time for the omci responses and the timers might fire
after deleting the device.

Change-Id: Ie41689194d944d533e118a24297b0a9685ade717